Update TEVE.txt - 6/11 just before huge update
tooltips for abilities in battle need to go away after clicked.

abilities on the ability panel in battle should be centered. so if we had 3 abilities the 2nd ability would be middle of the screen.

I made a .unitShadow but you can change it if needed. Please add this just below the unit image sprites for heroes and enemies in the battles. Remember to add it to the cleanup code as well.

tooltips for abilities in battle need to be larger, I'm thinking 24px font.
The exit battle button needs to match the formatting of the wave label in the top right.

leaving a battle is having issues when we return. we need to be sure that exiting a battle clears all variables of that battle so as not to interefere next time.

toggle auto is also having issues. we need a flag for the toggle at the start of a cycle and not check toggle again until that cycle ends.

on the heroes panel, we need the hero thumbnails at the bottom to be larger. I am thinking 140x140px. This means the font and the images will need to be larger as well.

It is difficult to see the stats on the hero info panel since the labels are far to the left and the stat values are far to the right. it would be nice to bring them much closer together. I'm thinking we do a double column, with HP, STR, AGI, INT on the left side with their values, and on the right side we put HP Regen, Attack Speed, Armor, and Resistance.

We also need the Back button to be slightly larger as well as the hero panel tab buttons text. 24px should be good for those.

I need the full changes posted. Give me the full index.html and battle.js files. You can't just make the changes and not show me. Either give me the full file if warranted, or show me the changes so I can manually make them. Don't change any code you don't have to. Everything is working so so well right now.